Install singularity with Ansible
It seems that Singularity cannot be installed with the packages
because it is in the "epel" repository which is not setup when
cloud-init is run. However the repository can be added through
dnf. So I added a new role which install software one by one.
So hopefully, I will be able to install Singularity on all nodes